# openapi.yaml
openapi: 3.0.0
info:
  title: Coffee Shop API
  description: API for the Coffee Shop Telegram WebApp
  version: 1.0.0
  contact:
    name: Coffee Shop Support
    url: https://coffeetelegram.website

servers:
  - url: https://coffeetelegram.website
    description: Production server

components:
  schemas:
    Coffee:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          maximum: 9223372036854775807
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
          minimum: 0
          maximum: 999999.99
        imageUrl:
          type: string

    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          maximum: 9223372036854775807
        userId:
          type: string
        status:
          type: string
          enum: [CREATED, PENDING, CONFIRMED, PREPARING, READY, COMPLETED, CANCELED]
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: float
          minimum: 0
          maximum: 999999.99
        createdAt:
          type: string
          format: date-time

    OrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          maximum: 9223372036854775807
        coffee:
          $ref: '#/components/schemas/Coffee'
        quantity:
          type: integer
          minimum: 1
          maximum: 100
        price:
          type: number
          format: float
          minimum: 0
          maximum: 999999.99

    CreateOrderRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - coffeeId
              - quantity
            properties:
              coffeeId:
                type: integer
                format: int64
                minimum: 1
                maximum: 9223372036854775807
              quantity:
                type: integer
                minimum: 1
                maximum: 100

paths:
  /api/coffee:
    get:
      summary: Get all available coffees
      description: Returns a list of all available coffee items
      responses:
        '200':
          description: List of coffee items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Coffee'
        '400':
          description: Bad Request - Invalid request parameters
        '401':
          description: Unauthorized - Missing or invalid Telegram authentication
        '406':
          description: Not Acceptable - Missing required header
        '500':
          description: Internal Server Error

  /api/coffee/{id}:
    get:
      summary: Get coffee by ID
      description: Returns a specific coffee item by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 9223372036854775807
      responses:
        '200':
          description: Coffee item found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coffee'
        '400':
          description: Bad Request - Invalid ID format
        '401':
          description: Unauthorized - Missing or invalid Telegram authentication
        '404':
          description: Coffee item not found
        '406':
          description: Not Acceptable - Missing required header
        '500':
          description: Internal Server Error

  /api/orders:
    post:
      summary: Create a new order
      description: Creates a new order with the specified items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Order created successfully
        '400':
          description: Bad Request - Invalid request data, empty items list or empty request body
        '401':
          description: Unauthorized - Missing or invalid Telegram authentication
        '404':
          description: Not Found - Coffee item not found
        '406':
          description: Not Acceptable - Missing required header
        '500':
          description: Internal Server Error

  /api/orders/user:
    get:
      summary: Get user's orders
      description: Returns all orders for the authenticated user
      responses:
        '200':
          description: List of user's orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Bad Request - Invalid request parameters
        '401':
          description: Unauthorized - Missing or invalid Telegram authentication
        '406':
          description: Not Acceptable - Missing required header
        '500':
          description: Internal Server Error

  /api/orders/{orderId}:
    delete:
      summary: Cancel an order
      description: Cancels a specific order if it belongs to the authenticated user
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 9223372036854775807
      responses:
        '200':
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad Request - Invalid ID format or order is already cancelled
        '401':
          description: Unauthorized - Missing or invalid Telegram authentication
        '403':
          description: Forbidden - Order does not belong to the user
        '404':
          description: Order not found
        '406':
          description: Not Acceptable - Missing required header
        '500':
          description: Internal Server Error 